services/rag-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
redis==5.0.8
requests==2.32.3
qdrant-client==1.10.1
fastembed
prometheus-client==0.20.0
pytest==8.3.4
pytest-cov==5.0.0   
httpx==0.27.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-httpx
opentelemetry-exporter-otlp